Initial commit
Server preparation playbook,

This playbook assumes that the machine is booted already. Also this should be
run as

$ ansible-playbook prepapre_server.yml -i ./configuration/inventory --ask-pass

and it will ask for the root password. After this playbook, ansible should have
passwordless ssh access.

This playbook will install sudo and wget, create a user exanest (with the same
password) and copy a public ssh key. It assumes that the host is running CentOS.